0720 로그인, 회원가입 로직 수정

+++ src/main/java/com/yp/puppy/api/service/user/UserService.java
@@ -50,6 +50,10 @@ public class UserService {
      * @return - true. 중복이 아닐 경우
      */
     public boolean checkEmailDuplicate(String email) {
+        if (email == null || email.isEmpty()) {
+            throw new IllegalArgumentException("이메일이 비어있습니다");
+        }
+
         boolean exists = userRepository.existsByEmail(email);
         log.info("Checking email {} is duplicated : {}", email, exists);
 
@@ -72,16 +76,21 @@ public class UserService {
      */
     private boolean notFinish(String email) {
         // 이메일로 이사람이 회원가입을 끝냈냐? 안끝냈냐 조회
-        User foundUser = userRepository.findByEmail(email).orElseThrow();
+        User foundUser = userRepository.findByEmail(email).orElse(null);
+        if (foundUser == null) {
+            // 유저가 존재하지 않으면 false 리턴
+            return false;
+        }
+
         if (!foundUser.isEmailVerified() || foundUser.getPassword() == null) {
-            // 이메일 인증 누락이거나              비번이 누락된 경우
+            // 이메일 인증 누락이거나 비번이 누락된 경우
             EmailVerification ev = emailVerificationRepository.findByUser(foundUser).orElse(null);
 
             if (ev != null) {
                 emailVerificationRepository.delete(ev);
             }
 
-            generateAndSendCode(email,foundUser);
+            generateAndSendCode(email, foundUser);
             return true;
         }
         return false;
@@ -213,6 +222,7 @@ public class UserService {
         String token = tokenProvider.createToken(user);
         return LoginResponseDto.builder()
                 .email(dto.getEmail())
+                .nickname(user.getNickname())
                 .role(user.getRole().toString())
                 .token(token)
                 .build();
